Extract packet ID generator
