+++ app/src/main/java/com/geeksville/mesh/service/PacketIdGenerator.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2025 Meshtastic LLC
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.geeksville.mesh.service
+
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import java.util.Random
+import javax.inject.Inject
+import kotlin.math.absoluteValue
+
+@InstallIn(SingletonComponent::class)
+class PacketIdGenerator @Inject constructor() {
+
+    // Do not use directly, instead call generatePacketId()
+    var currentPacketId = Random(System.currentTimeMillis()).nextLong().absoluteValue
+        private set
+
+    /** Generate a unique packet ID (if we know enough to do so - otherwise return 0 so the device will do it) */
+    @Suppress("MagicNumber")
+    @Synchronized
+    fun generate(): Int {
+        val numPacketIds = ((1L shl 32) - 1) // A mask for only the valid packet ID bits, either 255 or maxint
+
+        currentPacketId++
+
+        currentPacketId = currentPacketId and 0xffffffff // keep from exceeding 32 bits
+
+        // Use modulus and +1 to ensure we skip 0 on any values we return
+        return ((currentPacketId % numPacketIds) + 1L).toInt()
+    }
+}
